sentry logging and bug fixes
