+++ services/authorizer/workers/create_authorization/requirements.txt
@@ -1,3 +1,9 @@
+Flask==1.1.2
+Jinja2==2.11.2
+MarkupSafe==1.1.1
+Pygments==2.6.1
+SQLAlchemy==1.3.16
+Werkzeug==1.0.1
 appdirs==1.4.3
 appnope==0.1.0
 arrow==0.15.5
@@ -9,18 +15,14 @@ certifi==2020.4.5.1
 chardet==3.0.4
 click==7.1.2
 decorator==4.4.2
-Flask==1.1.2
-Flask-HTTPAuth==4.0.0
 idna==2.9
 ipdb==0.13.2
-ipython==7.14.0
 ipython-genutils==0.2.0
+ipython==7.14.0
 isort==4.3.21
 itsdangerous==1.1.0
 jedi==0.17.0
-Jinja2==2.11.2
 lazy-object-proxy==1.4.3
-MarkupSafe==1.1.1
 mccabe==0.6.1
 parso==0.7.0
 pathspec==0.8.0
@@ -30,18 +32,16 @@ pika==1.1.0
 prompt-toolkit==3.0.5
 psycopg2-binary==2.8.5
 ptyprocess==0.6.0
-Pygments==2.6.1
 pylint==2.5.0
 python-dateutil==2.8.1
 redis==3.5.0
 regex==2020.4.4
 requests==2.23.0
+sentry-sdk==0.14.4
 six==1.14.0
-SQLAlchemy==1.3.16
 toml==0.10.0
 traitlets==4.3.3
 typed-ast==1.4.1
 urllib3==1.25.9
 wcwidth==0.1.9
-Werkzeug==1.0.1
 wrapt==1.12.1
