+++ services/authorizer/workers/make_authorized_request/requirements.txt
@@ -1,3 +1,4 @@
+
 Flask-HTTPAuth==4.0.0
 Flask==1.1.2
 Jinja2==2.11.2
@@ -26,6 +27,7 @@ python-dateutil==2.8.1
 redis==3.5.0
 regex==2020.4.4
 requests==2.23.0
+sentry-sdk==0.14.4
 six==1.14.0
 toml==0.10.0
 typed-ast==1.4.1
