+++ services/authorizer/workers/make_authorized_request/worker.py
@@ -9,7 +9,12 @@ import os
 import pika
 import redis
 import requests
+import sentry_sdk
 import sys
+
+sentry_sdk.init(
+    "https://877d23fec9764314b6f0f15533ce1574@o398013.ingest.sentry.io/5253121"
+)
 
 current_module = sys.modules[__name__]
