+++ services/authorizer/workers/refresh_authorization/requirements.txt
@@ -23,6 +23,7 @@ python-dateutil==2.8.1
 redis==3.5.0
 regex==2020.4.4
 requests==2.23.0
+sentry-sdk==0.14.4
 six==1.14.0
 SQLAlchemy==1.3.16
 toml==0.10.0
