+++ services/player/requirements.txt
@@ -18,6 +18,7 @@ pathspec==0.8.0
 pylint==2.5.2
 redis==3.5.1
 regex==2020.5.7
+sentry-sdk[flask]==0.14.4
 six==1.14.0
 toml==0.10.0
 typed-ast==1.4.1
