+++ services/postgres/requirements.txt
@@ -10,6 +10,7 @@ pathspec==0.8.0
 psycopg2-binary==2.8.5
 pylint==2.5.0
 regex==2020.4.4
+sentry-sdk==0.14.4
 six==1.14.0
 toml==0.10.0
 typed-ast==1.4.1
