+++ services/room_settings/requirements.txt
@@ -25,6 +25,7 @@ pylint==2.5.0
 redis==3.5.0
 regex==2020.4.4
 requests==2.23.0
+sentry-sdk[flask]==0.14.4
 six==1.14.0
 toml==0.10.0
 typed-ast==1.4.1
