+++ services/room_settings/templates/index.html
@@ -26,7 +26,7 @@
 
 <body style='background-color:#191414; text-align:center;'>
 
-  <h1>{{ room_code }}></h1>
+  <h1>{{ room_code }}</h1>
   <p>*visit earbud.club and enter the room code*</p>
 </body>
