+++ services/typeahead/app.py
@@ -3,15 +3,21 @@
 from flask import Flask, request, jsonify
 from flask import redirect, make_response
 from flask_cors import CORS
-from requests.exceptions import HTTPError
 from functools import wraps
+from requests.exceptions import HTTPError
+from sentry_sdk.integrations.flask import FlaskIntegration
 import json
 import os
 import redis
+import sentry_sdk
 import spotify_request
 import sys
 import urllib
 
+sentry_sdk.init(
+    dsn="https://877d23fec9764314b6f0f15533ce1574@o398013.ingest.sentry.io/5253121",
+    integrations=[FlaskIntegration()],
+)
 
 current_module = sys.modules[__name__]
 
@@ -30,7 +36,9 @@ def spotify_search_response_to_universal_format(resp_dict):
         result_dict["type"] = track_dict["type"]
         result_dict["uri"] = track_dict["uri"]
         result_dict["id"] = track_dict["id"]
-        result_dict["album_art"] = track_dict["album"]["images"][0]["url"]
+        result_dict["album_art"] = next(iter(track_dict["album"]["images"]), {}).get(
+            "url", "https://shop.r10s.jp/hanjo/cabinet/c/imgrc0071257501.jpg"
+        )
         result_dict["artists"] = [a["name"] for a in track_dict["artists"]]
         results.append(result_dict)
     return results
