+++ services/typeahead/requirements.txt
@@ -22,6 +22,7 @@ pylint==2.5.2
 redis==3.5.1
 regex==2020.5.7
 requests==2.23.0
+sentry-sdk[flask]==0.14.4
 six==1.14.0
 toml==0.10.0
 typed-ast==1.4.1
