+++ tests/unit/authorization/workers/create_authorization/create_authorization_worker_test.py
@@ -8,3 +8,5 @@ import pytest
 @patch.object(services.player.app, "spotify_play_song")
 class TestPlayer(TestCase):
     def test_post_song_happy_path(self, play_song, am_send, redis_get):
+        pass
+
